FlurPilot-Core/apps/worker/pyproject.toml
[tool.poetry]
name = "flurpilot-worker"
version = "0.1.0"
description = "FlurPilot Core Worker (Hybrid Crawler & VPE)"
authors = ["The Core Team"]

[tool.poetry.dependencies]
python = "^3.12"
playwright = "^1.40.0"
geopandas = "^0.14.0"
redis = "^5.0.0"
pydantic = "^2.0.0"
openai = "^1.0.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
spacy = "^3.7.0"
pip-licenses = "^4.3.0"
# gVisor support implied by isolation strategy, not a python dependency

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
